# Optimize `set_key_and_attribute_body_from`