@byroot 2025-03-06
> by using `.equal?` for similar benefits.

Hum, but `equal?` doesn't have a dedicated opcode. So maybe just a `rubocop:disable`.

Let me know what you prefer.