@okuramasafumi 2025-03-06
@byroot Hi, thank you for your contribution. This is a really interesting optimization, although I do understand your explanation, I cannot come up with this by myself.
Ci is failing due to [`Style/YodaCondition` cop](https://docs.rubocop.org/rubocop/cops_style.html#styleyodacondition), which I believe should be turned off with `rubocop: disable` comment. In most cases the cop is still valid.